Replace shell scripts with integration test
