+++ .travis.yml
@@ -1,80 +1,26 @@
 language: rust
 
+branches:
+  only:
+    - master
+    - staging
+    - trying
+
+rust:
+  - 1.35.0
+  - stable
+  - nightly
+
+if: (branch = staging OR branch = trying) OR (type = pull_request AND branch = master)
+
 matrix:
   allow_failures:
     - rust: nightly
-  include:
-    - env: TARGET=x86_64-unknown-linux-gnu
-      rust: stable
-      if: (branch = staging OR branch = trying) OR (type = pull_request AND branch = master)
-
-    - env: TARGET=thumbv6m-none-eabi
-      rust: stable
-      if: (branch = staging OR branch = trying) OR (type = pull_request AND branch = master)
-
-    - env: TARGET=thumbv7m-none-eabi
-      rust: stable
-      if: (branch = staging OR branch = trying) OR (type = pull_request AND branch = master)
-
-    - env: TARGET=thumbv7em-none-eabi
-      rust: stable
-      if: (branch = staging OR branch = trying) OR (type = pull_request AND branch = master)
-
-    - env: TARGET=thumbv7em-none-eabihf
-      rust: stable
-      if: (branch = staging OR branch = trying) OR (type = pull_request AND branch = master)
-
-    # MSRV
-    - env: TARGET=thumbv6m-none-eabi
-      rust: 1.31.0
-      if: (branch = staging OR branch = trying) OR (type = pull_request AND branch = master)
-
-    # MSRV
-    - env: TARGET=thumbv7m-none-eabi
-      rust: 1.31.0
-      if: (branch = staging OR branch = trying) OR (type = pull_request AND branch = master)
-
-    # MSRV
-    - env: TARGET=thumbv7em-none-eabi
-      rust: 1.31.0
-      if: (branch = staging OR branch = trying) OR (type = pull_request AND branch = master)
-
-    # MSRV
-    - env: TARGET=thumbv7em-none-eabihf
-      rust: 1.31.0
-      if: (branch = staging OR branch = trying) OR (type = pull_request AND branch = master)
-
-    - env: TARGET=thumbv6m-none-eabi
-      rust: nightly
-      if: (branch = staging OR branch = trying) OR (type = pull_request AND branch = master)
-
-    - env: TARGET=thumbv7m-none-eabi
-      rust: nightly
-      if: (branch = staging OR branch = trying) OR (type = pull_request AND branch = master)
-
-    - env: TARGET=thumbv7em-none-eabi
-      rust: nightly
-      if: (branch = staging OR branch = trying) OR (type = pull_request AND branch = master)
-
-    - env: TARGET=thumbv7em-none-eabihf
-      rust: nightly
-      if: (branch = staging OR branch = trying) OR (type = pull_request AND branch = master)
-
-    - env: TARGET=thumbv8m.main-none-eabi
-      rust: nightly
-      if: (branch = staging OR branch = trying) OR (type = pull_request AND branch = master)
-
-    - env: TARGET=thumbv8m.base-none-eabi
-      rust: nightly
-      if: (branch = staging OR branch = trying) OR (type = pull_request AND branch = master)
 
 before_install: set -e
 
-install:
-  - bash ci/install.sh
-
 script:
-  - bash ci/script.sh
+  - cargo test --all
 
 after_script: set +e
 
@@ -82,12 +28,6 @@ cache: cargo
 
 before_cache:
   - chmod -R a+r $HOME/.cargo;
-
-branches:
-  only:
-    - master
-    - staging
-    - trying
 
 notifications:
   email:
