--- ci/install.sh
@@ -1,11 +1,0 @@
-set -euxo pipefail
-
-main() {
-    case $TARGET in
-        thumbv*-none-eabi*)
-            rustup target add $TARGET
-            ;;
-    esac
-}
-
-main
